Move batch system interface into a separate class instance.
Like we did earlier for the cloud provider interface, the
`Orchestrator` now *has an* instance of a batch system interface
(which is set at construction time), rather than *is a* batch system
interface (so that subclassing is necessary for changing the
interface).  This allows better composition of the VM-MAD pieces.

The new subpackage `vmmad.batchsys` hosts the standard batch system interfaces.

  new file:   batchsys/__init__.py
  modified:   orchestrator.py
  modified:   simul.py