New state `READY` for VMs.
The rationale is that the node name of a VM (and possibly public host
name, IP, etc.) might be known only after the VM has finished the boot
procedure and joined the cluster.

Therefore a new "notification" method `vm_is_ready` is implemented in
the `Orchestrator` class, to associate a VM instance with a node name
and pass it to the `READY` state.

  modified:   orchestrator.py
  modified:   simul.py